+++ meson.build
@@ -115,9 +115,10 @@ sources = [
   'src/game/gun/gun_pistols.c',
   'src/game/gun/gun_rifle.c',
   'src/game/input.c',
-  'src/game/inventry.c',
-  'src/game/invfunc.c',
-  'src/game/invvars.c',
+  'src/game/inventory/inventory_main.c',
+  'src/game/inventory/inventory_func.c',
+  'src/game/inventory/inventory_ring.c',
+  'src/game/inventory/inventory_vars.c',
   'src/game/items.c',
   'src/game/lara/lara_cheat.c',
   'src/game/lara/lara_col.c',
