+++ src/game/gameflow.c
@@ -6,7 +6,8 @@
 #include "game/fmv.h"
 #include "game/game.h"
 #include "game/input.h"
-#include "game/inv.h"
+#include "game/inventory/inventory_func.h"
+#include "game/inventory/inventory_vars.h"
 #include "game/lara/lara_main.h"
 #include "game/music.h"
 #include "game/output.h"
