+++ src/game/inventory/inventory_main.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdint.h>
+
+int32_t Inv_Display(int inv_mode);
