+++ src/game/objects/general/save_crystal.c
@@ -2,7 +2,7 @@
 
 #include "game/gameflow.h"
 #include "game/input.h"
-#include "game/inv.h"
+#include "game/inventory/inventory_main.h"
 #include "game/items.h"
 #include "game/lara/lara_main.h"
 #include "game/objects/general/pickup.h"
