+++ src/game/objects/general/scion.c
@@ -2,7 +2,7 @@
 
 #include "game/effects.h"
 #include "game/input.h"
-#include "game/inv.h"
+#include "game/inventory/inventory_func.h"
 #include "game/items.h"
 #include "game/lara/lara_main.h"
 #include "game/objects/common.h"
