+++ src/game/objects/traps/midas_touch.c
@@ -1,7 +1,8 @@
 #include "game/objects/traps/midas_touch.h"
 
 #include "game/input.h"
-#include "game/inv.h"
+#include "game/inventory/inventory_func.h"
+#include "game/inventory/inventory_main.h"
 #include "game/lara/lara_main.h"
 #include "game/objects/common.h"
 #include "global/const.h"
