+++ src/game/savegame_legacy.c
@@ -2,7 +2,7 @@
 #include "game/savegame_legacy.h"
 
 #include "game/gameflow.h"
-#include "game/inv.h"
+#include "game/inventory/inventory_func.h"
 #include "game/items.h"
 #include "game/lara/lara_main.h"
 #include "game/lot.h"
