+++ src/game/setup.c
@@ -3,7 +3,7 @@
 #include "config.h"
 #include "game/gamebuf.h"
 #include "game/gameflow.h"
-#include "game/inv.h"
+#include "game/inventory/inventory_main.h"
 #include "game/lara/lara_hair.h"
 #include "game/lara/lara_main.h"
 #include "game/level.h"
