+++ src/specific/s_input.c
@@ -1,7 +1,7 @@
 #include "specific/s_input.h"
 
 #include "config.h"
-#include "game/inv.h"
+#include "game/inventory/inventory_func.h"
 #include "game/lara/lara_main.h"
 #include "game/shell.h"
 #include "global/vars.h"
